+++ frontend/lib/screens/board/small_widget.dart
@@ -1,0 +1,20 @@
+
+
+import 'package:flutter/material.dart';
+
+class SmallWidget extends StatefulWidget {
+  const SmallWidget({super.key});
+
+  @override
+  SmallWidgetState createState() => SmallWidgetState();
+}
+
+class SmallWidgetState extends State<SmallWidget> {
+  @override
+  Widget build(BuildContext context) {
+    
+    return const Text("Small Widget.");
+
+  }
+
+}